% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprecruit.R
\name{cohort.props.f}
\alias{cohort.props.f}
\title{Proportion of a cohort recruiting each year}
\usage{
cohort.props.f(birth.year, final.year, Linf, k, t0, cv)
}
\arguments{
\item{birth.year}{four digit year of birth}

\item{final.year}{four digit final year of the growth projection (birth year + average max age is a good)}

\item{Linf}{VB maximum length (cm)}

\item{k}{VB k}

\item{t0}{VB t-zero}

\item{cv}{coefficient of variation on length for an age}

\item{lw.a}{a parameter for length (cm) - weight (g) exponential function}

\item{lw.b}{b parameter for length (cm) - weight (g) exponential function}
}
\description{
Using a Von Bertalanffy growth function and parameters you supply, this determines the proportion
of a cohort's numbers and biomass that are at or above different size classes each year
}
\examples{
cohort.props.f(birth.year=2011, final.year = 2030, Linf = 42, k = 0.086, t0 = -1.57, cv = 0.089)
vb.growth.f(age.vector=1:40,Linf=42,k=0.086,t0=-1.57,cv=0.1)
}
\keyword{VonBertalanffy}
\keyword{age}
\keyword{growth}
\keyword{recruit}
