% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprecruit.R
\name{propinterp.f}
\alias{propinterp.f}
\title{gam interpolation of sigmoidal curve}
\usage{
propinterp.f(recruiting.matrix, birth.year, final.year, len)
}
\arguments{
\item{birth.year}{four digit year of birth}

\item{final.year}{four digit final year of the growth projection (birth year + average max age is a good)}

\item{len}{length class of interest}

\item{proportion}{the proportion of interest}
}
\description{
Just for plotting. Gives the year where probability = 50%
}
