% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprecruit.R
\name{vb.growth.f}
\alias{vb.growth.f}
\title{Von Bertalanffy growth}
\usage{
vb.growth.f(age.vector, Linf, k, t0, cv)
}
\arguments{
\item{age.vector}{ages}

\item{Linf}{VB maximum length (cm)}

\item{k}{VB k}

\item{t0}{VB t-zero}

\item{cv}{coefficient of variation on length for an age}
}
\description{
Using a Von Bertalanffy growth function and parameters you supply, show the trajectory for the supplied age vector
}
\examples{
vb.growth.f(age.vector=1:40,Linf=42,k=0.086,t0=-1.57,cv=0.1)
cohort.props.f(birth.year=2011, final.year = 2030, Linf = 42, k = 0.086, t0 = -1.57, cv = 0.089)
}
\keyword{VonBertalanffy}
\keyword{age}
\keyword{growth}
\keyword{recruit}
