% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proprecruit.R
\name{VB.growth.for.age.f}
\alias{VB.growth.for.age.f}
\title{Von Bertalanffy growth solved for mean age given a length}
\usage{
VB.growth.for.age.f(length.vector, k, Linf, t0)
}
\arguments{
\item{length.vector}{ages}

\item{k}{VB k}

\item{Linf}{VB maximum length (cm)}

\item{t0}{VB t-zero}
}
\description{
Using a Von Bertalanffy growth function and parameters you supply, show the mean age given a specified length
}
\keyword{VonBertalanffy}
\keyword{age}
\keyword{growth}
\keyword{recruit}
